cost functions implemented? 2 heuristic functions implemented
some functions on operators.h need to be replaced in another file...

+++ node.h
@@ -30,7 +30,7 @@ class point {
 
 class node {
  public:
-  int cost;
+  double cost;
   std::vector<model> map;
   std::vector<point> pos;
   point target;
